requirements.txt
Flask
python-docx
google-generativeai
python-dotenv
langchain_google_genai
